Fix float32 precision bug causing bucket values to exceed 9999 for users with hash values near max uint32

+++ pkg/decision/bucketer/murmurhashbucketer.go
@@ -27,7 +27,7 @@ import (
 	"github.com/twmb/murmur3"
 )
 
-var maxHashValue = float64(math.Pow(2, 32))
+var maxHashValue = math.Pow(2, 32)
 
 // DefaultHashSeed is the hash seed to use for murmurhash
 const DefaultHashSeed = 1
